JPAR-99 - roll back version of Maven dependency plugin
The current version 3.0.2 (required for Java 9 compatibility) appears
to have a bug causing it to ignore the type of the artifact to be
unpacked in the unpack goal, so that a test-jar artifact cannot be
used. Rolling back the version number to 2.10 (the most recent version
not to have this problem) for now.